Moved ParameterType to its own file
