+++ lluvia/cpp/core/include/lluvia/core.h
@@ -37,6 +37,8 @@
 #include "core/node/NodeBuilderDescriptor.h"
 #include "core/node/NodeState.h"
 #include "core/node/NodeType.h"
+#include "core/node/Parameter.h"
+#include "core/node/ParameterType.h"
 #include "core/node/PortDescriptor.h"
 #include "core/node/PortDirection.h"
 #include "core/node/PortType.h"
@@ -46,7 +48,6 @@
 #include "core/Duration.h"
 #include "core/error.h"
 #include "core/Interpreter.h"
-#include "core/Parameter.h"
 #include "core/Program.h"
 #include "core/Session.h"
 #include "core/types.h"
