+++ lluvia/cpp/core/src/Interpreter.cpp
@@ -31,10 +31,10 @@
 #include "lluvia/core/node/ContainerNodeDescriptor.h"
 #include "lluvia/core/node/Node.h"
 #include "lluvia/core/node/NodeBuilderDescriptor.h"
+#include "lluvia/core/node/Parameter.h"
 #include "lluvia/core/node/PushConstants.h"
 
 #include "lluvia/core/Object.h"
-#include "lluvia/core/Parameter.h"
 #include "lluvia/core/Program.h"
 #include "lluvia/core/Session.h"
 #include "lluvia/core/types.h"
