+++ lluvia/python/src/BUILD.bazel
@@ -13,8 +13,6 @@ pyx_library (
         "lluvia/core/core_object.pxd",
         "lluvia/core/duration.pxd",
         "lluvia/core/duration.pyx",
-        "lluvia/core/parameter.pxd",
-        "lluvia/core/parameter.pyx",
         "lluvia/core/program.pxd",
         "lluvia/core/program.pyx",
         "lluvia/core/session.pxd",
