--- lluvia/python/src/lluvia/core/enums/parameter.pyx
@@ -1,9 +1,0 @@
-# cython: language_level=3, boundscheck=False, emit_code_comments=True, embedsignature=True
-
-"""
-    lluvia.core.enums.parameter
-    ---------------------------
-
-    :copyright: 2018, Juan David Adarve Bermudez. See AUTHORS for more details.
-    :license: Apache-2 license, see LICENSE for more details.
-"""
