+++ lluvia/python/src/lluvia/core/node/compute_node.pxd
@@ -13,13 +13,13 @@ from libcpp.string cimport string
 
 from lluvia.core.command_buffer cimport _CommandBuffer
 from lluvia.core.core_object cimport _Object
-from lluvia.core.parameter cimport _Parameter
 from lluvia.core.program cimport _Program
 from lluvia.core.session cimport Session
 from lluvia.core.types cimport _vec3ui
 
 from lluvia.core.node.compute_node_descriptor cimport _ComputeNodeDescriptor
 from lluvia.core.node.node_type cimport _NodeType
+from lluvia.core.node.parameter cimport _Parameter
 from lluvia.core.node.push_constants cimport _PushConstants
 
 cdef extern from 'lluvia/core/node/ComputeNode.h' namespace 'll':
