+++ lluvia/python/src/lluvia/core/node/container_node.pxd
@@ -11,12 +11,12 @@ from libcpp.string cimport string
 
 from lluvia.core.command_buffer cimport _CommandBuffer
 from lluvia.core.core_object cimport _Object
-from lluvia.core.parameter cimport _Parameter
 from lluvia.core.session cimport Session
 
-from lluvia.core.node.node_type cimport _NodeType
-from lluvia.core.node.node cimport _Node
 from lluvia.core.node.container_node_descriptor cimport _ContainerNodeDescriptor
+from lluvia.core.node.node cimport _Node
+from lluvia.core.node.node_type cimport _NodeType
+from lluvia.core.node.parameter cimport _Parameter
 
 cdef extern from 'lluvia/core/node/ContainerNode.h' namespace 'll':
