+++ lluvia/python/src/lluvia/core/node/container_node_descriptor.pxd
@@ -8,7 +8,7 @@
 
 from libcpp.string cimport string
 
-from lluvia.core.parameter cimport _Parameter
+from lluvia.core.node.parameter cimport _Parameter
 from lluvia.core.node.port_descriptor cimport _PortDescriptor
