+++ lluvia/python/src/lluvia/core/node/parameter_type.pyx
@@ -1,0 +1,11 @@
+"""
+    lluvia.core.node.parameter_type
+    -------------------------------
+
+    :copyright: 2022, Juan David Adarve Bermudez. See AUTHORS for more details.
+    :license: Apache-2 license, see LICENSE for more details.
+"""
+
+__all__ = [
+    'ParameterType'
+]
